sel4_panicking

Struct PanicMessage

pub struct PanicMessage<'a> { /* private fields */ }

A message that was given to the panic!() macro.

The [Display] implementation of this type will format the message with the arguments that were given to the panic!() macro.

See [PanicInfo::message].

Implementations

impl<'a> PanicMessage<'a>

pub const fn as_str(&self) -> Option<&'static str>

Gets the formatted message, if it has no arguments to be formatted at runtime.

This can be used to avoid allocations in some cases.

Guarantees

For panic!("just a literal"), this function is guaranteed to return Some("just a literal").

For most cases with placeholders, this function will return None.

See [fmt::Arguments::as_str] for details.